+++ FQHE/src/HilbertSpace/FermionOnSphereWithSpinSqueezedBasis.cc
@@ -1619,7 +1619,7 @@ double FermionOnSphereWithSpinSqueezedBasis::ComputeSign(unsigned long signs)
 // totalSpin = twce the total spin value
 // return value = Hilbert space dimension
 
-int FermionOnSphereWithSpinSqueezedBasis::EvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz, int totalSpin)
+long FermionOnSphereWithSpinSqueezedBasis::EvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz, int totalSpin)
 {
   //  codage des etats sur deux bits, -lzMax up down on the lsb's
   
@@ -1629,7 +1629,7 @@ int FermionOnSphereWithSpinSqueezedBasis::EvaluateHilbertSpaceDimension(int nbrF
   unsigned long i, coeff;
   int k, position; 
   int CheckLz;
-  int counter;
+  long counter;
   int DimOrbit = lzMax+1;
   /*-------------INITS---------------------*/
