+++ FQHE/src/HilbertSpace/FermionOnSphereWithSpinSqueezedBasis.h
@@ -279,7 +279,7 @@ class FermionOnSphereWithSpinSqueezedBasis :  public FermionOnSphereWithSpin
   // totalLz = momentum total value
   // totalSpin = twce the total spin value
   // return value = Hilbert space dimension
-  virtual int EvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz, int totalSpin);
+  virtual long EvaluateHilbertSpaceDimension(int nbrFermions, int lzMax, int totalLz, int totalSpin);
 
   // evaluate Hilbert space dimension
   //
